chore(): move test to test folder in pg-protocol
