+++ packages/pg-protocol/package.json
@@ -16,11 +16,10 @@
     "typescript": "^4.0.3"
   },
   "scripts": {
-    "test": "mocha dist/**/*.test.js",
-    "build": "tsc",
-    "build:watch": "tsc --watch",
-    "prepublish": "yarn build",
-    "pretest": "yarn build"
+    "build": "tsc --build",
+    "build:watch": "tsc --build --watch",
+    "test": "mocha -r ts-node/register test/**/*.ts",
+    "prepublish": "yarn build"
   },
   "repository": {
     "type": "git",
