+++ packages/pg-protocol/test/inbound-parser.test.ts
@@ -1,9 +1,9 @@
 import buffers from './testing/test-buffers'
 import BufferList from './testing/buffer-list'
-import { parse } from '.'
+import { parse } from '../src'
 import assert from 'assert'
 import { PassThrough } from 'stream'
-import { BackendMessage } from './messages'
+import { BackendMessage } from '../src/messages'
 
 var authOkBuffer = buffers.authenticationOk()
 var paramStatusBuffer = buffers.parameterStatus('client_encoding', 'UTF8')
