+++ packages/pg-protocol/tsconfig.json
@@ -11,7 +11,11 @@
     "outDir": "dist",
     "incremental": true,
     "baseUrl": ".",
-    "declaration": true
+    "declaration": true,
+    "types": [
+      "mocha",
+      "node"
+    ]
   },
   "include": [
     "src/**/*"
